+++ servers/dsub/jobs/controllers/jobs_controller.py
@@ -207,6 +207,14 @@ def get_task_attempts(id, task, **kwargs):
     raise NotImplementedError('Tasks not supported by dsub.')
 
 
+def get_operation_details(job, operation, **kwargs):
+    """
+    Placeholder for query for operation details from Google Pipelines API
+
+    """
+    raise NotImplementedError('Operation details not supported by dsub.')
+
+
 def _handle_http_error(error, proj_id):
     # TODO(https://github.com/googlegenomics/dsub/issues/79): Push this
     # provider-specific error translation down into dstat.